clean weather code done
