+++ app/src/main/java/kpfu/itis/valisheva/android_app/data/response/SystemDescription.kt
@@ -3,8 +3,6 @@ package kpfu.itis.valisheva.android_app.data.response
 import com.google.gson.annotations.SerializedName
 
 data class SystemDescription(
-//    @SerializedName("id")
-//    val id: Int,
     @SerializedName("country")
     val country: String,
 )
