--- app/src/main/java/kpfu/itis/valisheva/android_app/data/response/WeatherShortInformation.kt
@@ -1,10 +1,0 @@
-package kpfu.itis.valisheva.android_app.data.response
-
-import com.google.gson.annotations.SerializedName
-
-data class WeatherShortInformation(
-    @SerializedName("id")
-    val id :Int,
-    @SerializedName("main")
-    val baseDescription: BaseDescription,
-)
