+++ app/src/main/java/kpfu/itis/valisheva/android_app/data/response/main/NearWeatherResponse.kt
@@ -1,0 +1,11 @@
+package kpfu.itis.valisheva.android_app.data.response.main
+
+import com.google.gson.annotations.SerializedName
+
+data class NearWeatherResponse(
+    @SerializedName("cod")
+    val statusCode: Int,
+    @SerializedName("list")
+    val cities : ArrayList<WeatherResponse>
+
+)
