+++ app/src/main/java/kpfu/itis/valisheva/android_app/data/response/main/WeatherResponse.kt
@@ -1,0 +1,29 @@
+package kpfu.itis.valisheva.android_app.data.response.main
+
+import com.google.gson.annotations.SerializedName
+import kpfu.itis.valisheva.android_app.data.response.*
+
+data class WeatherResponse(
+    @SerializedName("id")
+    val id :Int,
+    @SerializedName("coord")
+    val coordinates: Coordinates,
+    @SerializedName("weather")
+    val weather: ArrayList<Weather>,
+    @SerializedName("main")
+    val baseDescription: BaseDescription,
+    @SerializedName("wind")
+    val wind: Wind,
+    @SerializedName("clouds")
+    val cloudsPercent: Clouds,
+    @SerializedName("sys")
+    val systemDescription: SystemDescription,
+    @SerializedName("name")
+    val name: String,
+    @SerializedName("cod")
+    val statusCod: Int,
+
+    )
+
+
+
