+++ app/src/main/java/kpfu/itis/valisheva/android_app/data/response/main/WeatherShortResponse.kt
@@ -1,0 +1,8 @@
+package kpfu.itis.valisheva.android_app.data.response.main
+
+import com.google.gson.annotations.SerializedName
+
+data class WeatherShortResponse(
+    @SerializedName("id")
+    val id :Int,
+)
