+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/entities/City.kt
@@ -1,0 +1,5 @@
+package kpfu.itis.valisheva.android_app.domain.entities
+
+data class City(
+    val id: Int,
+)
