+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/entities/Coordinates.kt
@@ -1,0 +1,7 @@
+package kpfu.itis.valisheva.android_app.domain.entities
+
+data class Coordinates(
+    val latitude: Double,
+    val longitude: Double,
+)
+
