+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/entities/ShortCityWeather.kt
@@ -1,0 +1,7 @@
+package kpfu.itis.valisheva.android_app.domain.entities
+
+data class ShortCityWeather(
+    val id: Int,
+    val name: String,
+    val temp: Double,
+)
