+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/exceptions/NotFoundLocationException.kt
@@ -1,0 +1,6 @@
+package kpfu.itis.valisheva.android_app.domain.exceptions
+
+import java.lang.RuntimeException
+
+class NotFoundLocationException(message: String?, cause: Throwable?) :
+    RuntimeException(message, cause)
