+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/helpers/NavigationConstants.kt
@@ -1,0 +1,20 @@
+package kpfu.itis.valisheva.android_app.domain.helpers
+
+object NavigationConstants {
+    const val WEST = "West"
+    const val SOUTH = "South"
+    const val NORTH = "North"
+    const val EAST = "East"
+    const val NE = "NE"
+    const val SE = "SE"
+    const val SW = "SW"
+    const val NW = "NW"
+    const val EEN = "EEN"
+    const val NNE = "NNE"
+    const val ESE = "ESE"
+    const val SSE = "SSE"
+    const val SSW = "SSW"
+    const val WSW = "WSW"
+    const val WNW = "WNW"
+    const val NWN = "NNW"
+}
