+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/repositories/LocationRepository.kt
@@ -1,0 +1,9 @@
+package kpfu.itis.valisheva.android_app.domain.repositories
+
+import kpfu.itis.valisheva.android_app.domain.entities.Coordinates
+
+interface LocationRepository {
+
+    fun getLocation(): Coordinates
+    fun getDefaultLocation(): Coordinates
+}
