+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/usecases/location/GetLocationUseCase.kt
@@ -1,0 +1,17 @@
+package kpfu.itis.valisheva.android_app.domain.usecases.location
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import kpfu.itis.valisheva.android_app.domain.entities.Coordinates
+import kpfu.itis.valisheva.android_app.domain.entities.ShortCityWeather
+import kpfu.itis.valisheva.android_app.domain.repositories.LocationRepository
+import kpfu.itis.valisheva.android_app.domain.repositories.WeatherRepository
+
+class GetLocationUseCase(
+    private val locationRepository: LocationRepository
+){
+    operator fun invoke(): Coordinates {
+        return locationRepository.getLocation()
+    }
+}
