+++ app/src/main/java/kpfu/itis/valisheva/android_app/domain/usecases/weather/GetWeatherUseCase.kt
@@ -1,0 +1,18 @@
+package kpfu.itis.valisheva.android_app.domain.usecases.weather
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import kpfu.itis.valisheva.android_app.domain.entities.City
+import kpfu.itis.valisheva.android_app.domain.repositories.WeatherRepository
+
+class GetWeatherUseCase(
+    private val weatherRepository: WeatherRepository,
+    private val dispatcher: CoroutineDispatcher = Dispatchers.IO
+) {
+    suspend operator fun invoke(name: String): City{
+        return withContext(dispatcher){
+            weatherRepository.getWeather(name)
+        }
+    }
+}
