--- app/src/main/java/kpfu/itis/valisheva/android_app/exceptions/NotFoundLocationException.kt
@@ -1,6 +1,0 @@
-package kpfu.itis.valisheva.android_app.exceptions
-
-import java.lang.RuntimeException
-
-class NotFoundLocationException(message: String?, cause: Throwable?) :
-    RuntimeException(message, cause)
