--- app/src/main/java/kpfu/itis/valisheva/android_app/models/City.kt
@@ -1,7 +1,0 @@
-package kpfu.itis.valisheva.android_app.models
-
-data class City(
-    val id: Int,
-    val name: String,
-    val temp: Double,
-)
