+++ app/src/main/java/kpfu/itis/valisheva/android_app/presentation/extentions/ActivityExt.kt
@@ -1,0 +1,9 @@
+package kpfu.itis.valisheva.android_app.presentation.extentions
+
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.fragment.NavHostFragment
+
+fun AppCompatActivity.findController(
+    id : Int
+) = (supportFragmentManager.findFragmentById(id) as NavHostFragment)
+    .navController
